+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## 5.0.0-beta.0 (2018-04-06)
+## 5.0.0-beta.1 (2018-04-06)
 
 ### Features
 
 - `catchOffline` RxJS operator
 - `canActivate`, `canActivateChild` and `canLoad` guards
+- `isOnline` to check online status
+- `connectionChanges` to observe when online status changes